// client/src/pages/admin/types.ts
export enum NavMenuItemEnum {
  PointTable = 'point-table',
  PuzzleSetting = 'puzzle-setting',
  PuzzleStatus = 'puzzle-status',
  Reset = 'reset',
  Statistics = 'statistics',
  TeamPassword = 'team-password',
  TeamPoint = 'team-point',
  Timer = 'timer',
  Upload = 'upload',
  AdminPassword = 'admin-password',
}